requirements.txt
Django>=5.0,<6
django-allauth>=64.0
python-dotenv>=1.0
